Extract cost tracking to cost.py, remove static pricing table
